Updated the User Interface to make it user friendly. Also updated the delete functionality now confirmation is required to delete any tasbeeh.

+++ tasbeeh/lib/TasbeehCounter.dart
@@ -1,3 +1,5 @@
+import 'dart:ui';
+
 import 'package:flutter/material.dart';
 import 'DataBaseHelper.dart';
 
@@ -80,8 +82,40 @@ class TasbeehCounterState extends State<TasbeehCounter> {
             ),
             InkWell(
               onTap: () async {
-                await DatabaseHelper().deleteTasbeeh(widget.id);
-                Navigator.pop(context, true);
+                bool confirmDelete = await showDialog(
+                  context: context,
+                  builder: (context) => AlertDialog(
+                    backgroundColor: Colors.red, // Set background color to red for danger
+                    title: const Row(
+                      children: [
+                        Icon(Icons.warning, color: Colors.yellow), // Add warning icon
+                        SizedBox(width: 10),
+                        Text("Confirm Deletion", style: TextStyle(color: Colors.white)),
+                      ],
+                    ),
+                    content: const Text("Are you sure you want to delete this Tasbeeh?", style: TextStyle(color: Colors.white)),
+                    actions: [
+                      TextButton(
+                        onPressed: () {
+                          // Dismiss the dialog and confirm deletion
+                          Navigator.of(context).pop(true);
+                        },
+                        child: const Text("Yes", style: TextStyle(color: Colors.white)),
+                      ),
+                      TextButton(
+                        onPressed: () {
+                          // Dismiss the dialog and cancel deletion
+                          Navigator.of(context).pop(false);
+                        },
+                        child: const Text("No", style: TextStyle(color: Colors.white)),
+                      ),
+                    ],
+                  ),
+                );
+                if(confirmDelete){
+                  await DatabaseHelper().deleteTasbeeh(widget.id);
+                  Navigator.pop(context, true);
+                }
               },
               child: const Icon(Icons.delete),
             )
@@ -90,96 +124,99 @@ class TasbeehCounterState extends State<TasbeehCounter> {
       ),
       body: SingleChildScrollView(
         scrollDirection: Axis.vertical,
-        child: Center(
-          child: Container(
-            color: Colors.white,
-            child: Column(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          crossAxisAlignment: CrossAxisAlignment.center,
+          children: [
+            const SizedBox(height: 20),
+            Row(
               mainAxisAlignment: MainAxisAlignment.center,
-              children: <Widget>[
-                Image.asset('assets/images/masjid.webp'),
-                Padding(
-                  padding: const EdgeInsets.all(16.0),
-                  child: Column(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: [
-                      Text(
-                        widget.tasbeeh,
-                        textAlign: TextAlign.center,
-                        style: const TextStyle(
-                          fontSize: 24,
-                          fontWeight: FontWeight.w900,
-                          color: Colors.red,
-                        ),
-                      ),
-                      const SizedBox(height: 10),
-                      Text(
-                        'Total Tasbeeh: $_cycles',
-                        style: const TextStyle(
-                          color: Colors.deepPurple,
-                          fontSize: 20,
-                          fontWeight: FontWeight.w900,
-                        ),
-                      ),
-                      const SizedBox(height: 10),
-                      const Text(
-                        'You have read your tasbeeh this many time(s):',
-                        textAlign: TextAlign.center,
-                        style: TextStyle(
-                          fontSize: 24,
-                          fontWeight: FontWeight.w900,
-                          color: Colors.deepPurple,
-                        ),
-                      ),
-                      const SizedBox(height: 10),
-                      Text(
-                        '$_counter',
-                        style: const TextStyle(
-                          color: Colors.deepPurple,
-                          fontSize: 100,
-                          fontWeight: FontWeight.w900,
-                        ),
-                      ),
-                      const SizedBox(height: 20),
-                      ElevatedButton(
-                        onPressed: _incrementCounter,
-                        style: ElevatedButton.styleFrom(
-                          backgroundColor: Colors.white.withOpacity(0.3),
-                          shape: const CircleBorder(),
-                          padding: const EdgeInsets.all(70),
-                          side: const BorderSide(color: Colors.black, width: 3),
-                        ),
-                        child: const Text(
-                          'Add Count',
-                          style: TextStyle(
-                            color: Colors.black,
-                            fontSize: 30,
-                            fontWeight: FontWeight.w900,
-                          ),
-                        ),
-                      ),
-                      const SizedBox(height: 20), // Added some space between buttons
-                      ElevatedButton(
-                        onPressed: _resetCounter,
-                        style: ElevatedButton.styleFrom(
-                          backgroundColor: Colors.white.withOpacity(0.3),
-                          side: const BorderSide(color: Colors.black, width: 3),
-                        ),
-                        child: const Text(
-                          'Reset Count',
-                          style: TextStyle(
-                            color: Colors.black,
-                            fontSize: 20,
-                            fontWeight: FontWeight.w900,
-                          ),
-                        ),
-                      ),
-                      const SizedBox(height: 20), // Added some space at the bottom
-                    ],
+              children: [
+                const Text('Tasbeeh: ',
+                  style: TextStyle(
+                      fontSize: 24,
+                      fontWeight: FontWeight.w900,
+                      color: Colors.deepPurple,
+                  ),
+                ),
+                Text(widget.tasbeeh,
+                  style: const TextStyle(
+                    fontSize: 24,
+                    fontWeight: FontWeight.w900,
+                    color: Colors.brown,
+                    decoration: TextDecoration.underline,
+                    decorationThickness: 2
                   ),
                 ),
               ],
             ),
-          ),
+            const SizedBox(height: 10),
+            Text(
+              'You have read your tasbeeh this many time(s): $_cycles',
+              textAlign: TextAlign.center,
+              style: const TextStyle(
+                color: Colors.deepPurple,
+                fontSize: 24,
+                fontWeight: FontWeight.w900,
+              ),
+            ),
+            const SizedBox(height: 10),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                Text(
+                  '$_counter/',
+                  style: const TextStyle(
+                    color: Colors.deepPurple,
+                    fontSize: 100,
+                    fontWeight: FontWeight.w900,
+                  ),
+                ),
+                Text('${widget.maximumCount}',
+                  style: const TextStyle(
+                    color: Colors.deepPurple,
+                    fontSize: 30,
+                    fontWeight: FontWeight.w900,
+                  ),
+                )
+              ],
+            ),
+            const SizedBox(height: 20),
+            ElevatedButton(
+              onPressed: _incrementCounter,
+              style: ElevatedButton.styleFrom(
+                backgroundColor: Colors.white.withOpacity(0.3),
+                shape: const CircleBorder(),
+                padding: const EdgeInsets.all(70),
+                side: const BorderSide(color: Colors.black, width: 3),
+              ),
+              child: const Text(
+                'Add Count',
+                style: TextStyle(
+                  color: Colors.black,
+                  fontSize: 30,
+                  fontWeight: FontWeight.w900,
+                ),
+              ),
+            ),
+            const SizedBox(height: 20), // Added some space between buttons
+            ElevatedButton(
+              onPressed: _resetCounter,
+              style: ElevatedButton.styleFrom(
+                backgroundColor: Colors.white.withOpacity(0.3),
+                side: const BorderSide(color: Colors.black, width: 3),
+              ),
+              child: const Text(
+                'Reset Count',
+                style: TextStyle(
+                  color: Colors.black,
+                  fontSize: 20,
+                  fontWeight: FontWeight.w900,
+                ),
+              ),
+            ),
+            const SizedBox(height: 20), // Added some space at the bottom
+          ],
         ),
       ),
     );
